feat: change blockchain JSON-RPC and CLI to limit by block number
Instead of limiting by epoch number

+++ src/cli/node/json_rpc_client.rs
@@ -42,7 +42,7 @@ pub fn raw(addr: SocketAddr) -> Result<(), failure::Error> {
     }
 }
 
-pub fn get_blockchain(addr: SocketAddr, epoch: i64, limit: u32) -> Result<(), failure::Error> {
+pub fn get_blockchain(addr: SocketAddr, epoch: i64, limit: i64) -> Result<(), failure::Error> {
     let mut stream = start_client(addr)?;
     let params = GetBlockChainParams { epoch, limit };
     let response = send_request(
